Bill S4313 version original - Luis R. Sepúlveda
BILL NUMBER: S4313 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the criminal procedure law, in relation to requiring all police officers and peace officers involved in the execution of a search warrant or other entry onto a premises to wear body cameras PURPOSE: To protect the safety and rights of individuals and increase police accountability when during the execution of a search warrant or entering a premises. SUMMARY OF PROVISIONS: Section one of the bill adds subdivision 3-a to § 690.