Bill S4314A version original - Luis R. Sepúlveda
BILL NUMBER: S4314 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the agriculture and markets law, in relation to using the term "human food grade" for certain pet food PURPOSE: To create a voluntary standard for pet foods to use the term "human food grade" on their labeling, and to prohibit pet foods that do not meet the standard from using the term and other misleading words or images. SUMMARY OF PROVISIONS: Section 1 adds a new Ag & Mkts L. § 129-b.