Bill S4292 version original - Andrew J. Lanza
BILL NUMBER: S4292 SPONSOR: LANZA TITLE OF BILL: An act to amend the real property tax law, in relation to eligibility for the disabled homeowners' exemption for a surviving spouse of a deceased person with a disability PURPOSE OR GENERAL IDEA OF THE BILL: Expands eligibility for the disabled homeowners' exemption to certain surviving spouses. SUMMARY OF SPECIFIC PROVISIONS: Section 1.