Bill S4291 version original - Andrew J. Lanza
BILL NUMBER: S4291 SPONSOR: LANZA TITLE OF BILL: An act to amend the penal law, in relation to the crime of falsely accusing someone of a crime PURPOSE: Creates the crime of falsely accusing someone of a crime. Makes such accusations a Class E Felony. SUMMARY OF PROVISIONS: Section one adds a new section 240.49 to the penal law. Section two is an effective date of November 1. JUSTIFICATION: This bill creates the crime of falsely accusing someone of a crime.