Bill S4333 version original - Luis R. Sepúlveda
BILL NUMBER: S4333 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the correction law, in relation to amending the term residential treatment facility; and to repeal subdivision 10 of section 73 of the correction law relating thereto PURPOSE: To stop the practice of having people paroled to a unit inside a correc- tional facility. SUMMARY OF PROVISIONS: Section 1 amends the definition of "residential treatment facility". Section 2 repeals Correction Law § 73(10).