Bill S4293 version original - Andrew J. Lanza
BILL NUMBER: S4293 SPONSOR: LANZA TITLE OF BILL: An act to amend the New York city charter, in relation to the applica- tion of uniform land use review procedures PURPOSE: This bill provides that any not-for-profit organization or corporation which provides or intends to provide shelter to homeless, be subject to local/municipal land use community review procedures. SUMMARY OF PROVISIONS: Applications for such land use shall be reviewed pursuant to a uniform review procedure.