Bill S4296 version original - Andrew J. Lanza
BILL NUMBER: S4296 SPONSOR: LANZA TITLE OF BILL: An act to amend the correction law, in relation to information contained in the sex offender registry PURPOSE: Requires the Division of Criminal Justice Services' (DCJS) website that provides sex offender registry information to be searchable by the zip code of a registrant's employment. SUMMARY OF PROVISIONS: Section one amends subdivision 1 of section 168-q of the correction law as amended by chapter 532 of the laws of 2011.