Bill S4321 version original - Andrew J. Lanza
BILL NUMBER: S4321 SPONSOR: LANZA TITLE OF BILL: An act to amend the mental hygiene law, in relation to the distribution in counties with the most prevalent prescription opioid abuse of educa- tional materials regarding the misuse of and addiction to prescription drugs PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to help to prevent the misuse or abuse of prescription drugs. SUMMARY OF SPECIFIC PROVISIONS: Amends 19.