Bill S4338 version original - Luis R. Sepúlveda
BILL NUMBER: S4338 SPONSOR: SEPULVEDA TITLE OF BILL: An act relating to establishing a joint legislative committee on racial equity in New York state PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to provide a joint racial equity impact assessment on all legislation that is favorable to New York state. SUMMARY OF PROVISIONS: Section I a. There is hereby created the joint legislative committee on racial equity in New York State.