Bill S4325 version original - Andrew J. Lanza
BILL NUMBER: S4325 SPONSOR: LANZA TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to establishing distinctive "POW-MIA" license plates; and to amend the state finance law, in relation to establishing the "POW-MIA awareness fund" PURPOSE: To establish a distinctive POW-MIA license plates and establishing the "POW-MIA Awareness Fund". SUMMARY OF PROVISIONS: Section 1 amends the vehicle and traffic law by adding a new section 404-ff.