Bill S4254 version original - Jake Ashby
BILL NUMBER: S4254 SPONSOR: ASHBY TITLE OF BILL: An act to amend the state finance law, in relation to providing state reimbursement for lost tax revenue due to devaluation of land subject to toxic contamination PURPOSE OR GENERAL IDEA OF BILL: Provides for state reimbursement for lost tax revenue due to devaluation of land as a result of toxic contamination SUMMARY OF PROVISIONS: Section 1. Subdivision 2 of section 5 of the legislative law, as added by section 7 of part CC of Section 1.