Bill S4297 version original - Andrew J. Lanza
BILL NUMBER: S4297 SPONSOR: LANZA TITLE OF BILL: An act to amend the penal law, in relation to establishing the crime of larceny of a controlled substance PURPOSE OR GENERAL IDEA OF BILL: To establish the crime of larceny of a controlled substance. SUMMARY OF SPECIFIC PROVISIONS: Section 1 of the bill adds a new section to Penal law 220.