Bill S4294 version original - Andrew J. Lanza
BILL NUMBER: S4294 SPONSOR: LANZA TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to parking permit systems for park-and-rides on Staten Island PURPOSE: Relates to authorizing a parking permit system for park-and-rides in the Borough of Staten Island SUMMARY OF PROVISIONS: Section 1 amends the Vehicle and Traffic law by adding a new section 1640-q. JUSTIFICATION: This bill would implement a parking permit system for park-and-ride parking lots located in Staten Island.