Bill S4282 version original - Mark Walczyk
BILL NUMBER: S4282 SPONSOR: WALCZYK TITLE OF BILL: An act to amend the education law, the public health law and the civil practice law and rules, in relation to requiring applicants for certain nursing licenses and certifications to complete training in the prevention of human trafficking PURPOSE OR GENERAL IDEA OF BILL: To train applicants for nursing licenses on the issue of human traffick- ing prevention. SUMMARY OF PROVISIONS: Section 1 amends section 305 of the education law.