Bill S4277 version original - Andrew J. Lanza
BILL NUMBER: S4277 SPONSOR: LANZA TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to the defi- nitions of the terms "impaired" and "intoxication" for the purposes of such law PURPOSE: To codify accepted legal definitions and include alteration of physical or mental abilities by known or unknown substances. SUMMARY OF PROVISIONS: Section 119-b defines impaired in conformity with the language of the New York State Court of Appeals ruling in People v. Cruz, 48 NY.