Bill S4315 version original - Luis R. Sepúlveda
BILL NUMBER: S4315 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the correction law, in relation to establishing the family reunion program PURPOSE OR GENERAL IDEA OF BILL: To provide for a family reunion program in NY state prisons. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends the correction law by adding a new section 138-b to create a family reunion program. Section 2 is the effective date.