Bill S4288 version original - Jake Ashby
BILL NUMBER: S4288 SPONSOR: ASHBY TITLE OF BILL: An act to amend the parks, recreation and historic preservation law, in relation to allowing veterans with a service-connected disability or Purple Heart free use of parks and campsites PURPOSE OR GENERAL IDEA OF BILL: To provide free use of state parks and campsites for disabled veterans and recipients of the Purple Heart. SUMMARY OF PROVISIONS: Section 1 amends section 13.