Bill S4346 version original - Luis R. Sepúlveda
BILL NUMBER: S4346 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the penal law, in relation to requiring that a defendant must be over twenty-five years old to be guilty of felony murder PURPOSE OR GENERAL IDEA OF BILL: This bill will raise the age for a felony murder defendant to be over twenty-five years of age. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends Penal Law 125.25 (3). Section 2 sets forth the effective date.