Bill S4269A version original - Cordell Cleare
BILL NUMBER: S4269 SPONSOR: CLEARE TITLE OF BILL: An act to amend the civil rights law and the education law, in relation to establishing a cause of action for fertility fraud PURPOSE: To provide civil recourse to patients, spouses, and children when a. physician service provider, or donor commits fertility fraud. SUMMARY OF PROVISIONS: Section one adds a new section 52-d to the civil rights law, creating a private right of action for fertility fraud.