Bill S4451 version original - Liz Krueger
BILL NUMBER: S4451 SPONSOR: KRUEGER TITLE OF BILL: An act to amend the correction law, in relation to risk assessment instruments for sex offenders PURPOSE: To update the guidelines of the sex offender risk assessment instrument and require use of a validated instrument. SUMMARY OF PROVISIONS: Section 1 makes a technical amendment to section 168-d of the correction law requiring the application of a risk assessment instrument.