Bill S4384 version original - Mark Walczyk
BILL NUMBER: S4384 SPONSOR: WALCZYK TITLE OF BILL: An act to amend the criminal procedure law, in relation to permitting the electronic appearance of a defendant in the county of Jefferson PURPOSE OR GENERAL IDEA OF BILL: To amend the Criminal Procedure Law permitting the electronic appearance of defendants in Jefferson County. SUMMARY OF SPECIFIC PROVISIONS: Subdivision 1 of Section 182.