Bill S4394 version original - Liz Krueger
BILL NUMBER: S4394 SPONSOR: KRUEGER TITLE OF BILL: An act to amend the administrative code of the city of New York, in relation to the use of bicycles for commercial purposes PURPOSE: To provide that businesses using bicycles for commercial purposes in the city of New York shall be responsible for certain traffic infractions of bicycle operators riding on their behalf. Also to provide that motor-as- sisted bicycles are not permitted to be ridden on sidewalks in the city of New York.