Bill S4403 version original - Leroy Comrie
BILL NUMBER: S4403 SPONSOR: COMRIE TITLE OF BILL: An act to amend the workers' compensation law, in relation to providing family leave in the event of a stillbirth or miscarriage PURPOSE: To extend New York's Paid Family Leave Law to cover cases of stillbirth or miscarriage. SUMMARY OF PROVISIONS: Ensures that those who endure cases of stillbirth or miscarriage are afforded Paid Family Leave; of one to four weeks.