Bill S4399 version original - Monica Martinez
BILL NUMBER: S4399 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the public health law, in relation to the sale of flavored vapor products PURPOSE: The purpose of this bill is to strengthen the prohibition of the sale of flavored vapor products in the State of New York by closing a loophole exploited by retails and stores. SUMMARY OF PROVISIONS: Section 1.