Bill S4441 version original - Nathalia Fernandez
BILL NUMBER: S4441 SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the penal law and the civil rights law, in relation to prohibiting speech-based defenses to actions brought against an individ- ual for the unlawful dissemination or publication of an intimate image PURPOSE OR GENERAL IDEA OF BILL: Prohibit using freedom of speech as a defense in revenge porn. SUMMARY OF PROVISIONS: Section 1 amends the penal law to ban the use of free speech as a defense for revenge porn.