Bill S4434 version original - Nathalia Fernandez
BILL NUMBER: S4434 SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the domestic relations law and the family court act, in relation to child custody forensic reports PURPOSE OR GENERAL IDEA OF BILL: This legislation provides for uniform access to court ordered forensic mental health evaluation reports and underlying data by litigants, their counsels and the counsel for the attorney for the child in child custody and visitation cases. SUMMARY OF PROVISIONS: Sections 1 and 2.