Bill S4424 version original - Nathalia Fernandez
BILL NUMBER: S4424 SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the criminal procedure law, in relation to admissibility of victim's outcry and disclosures in cases of sexual abuse PURPOSE OR GENERAL IDEA OF BILL: To allow testimony of a victim's disclosure of sexual assault, known as an "outcry," regardless of when that disclosure was made relative to their abuse. SUMMARY OF PROVISIONS: Section 1 amends the Criminal Procedure Law by adding section 60.