Bill S4419 version original - Nathalia Fernandez
BILL NUMBER: S4419 SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the social services law, the executive law, and the public health law, in relation to providing for supports and services for unaccompanied minors with no lawful immigration status PURPOSE OR GENERAL IDEA OF BILL: This bill would provide supports and services for unaccompanied minors that are released to a sponsor in New York State. SUMMARY OF PROVISIONS: Section 1 provides a definition for an unaccompanied minor.