Bill S4421 version original - Nathalia Fernandez
BILL NUMBER: S4421 SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the criminal procedure law, in relation to permitting support to a vulnerable elderly person testifying in grand jury proceedings PURPOSE OR GENERAL IDEA OF BILL: To allow a social worker or informal caregiver to be present and give care to a vulnerable elderly person during grand jury proceedings. SUMMARY OF PROVISIONS: Section one amends paragraph (h) of subdivision 3 of § 190.