Bill S4426 version original - Nathalia Fernandez
BILL NUMBER: S4426 SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the penal law, in relation to modifying the offense of "forcible touching" to "unwanted touching" and by removing certain elements of the actor's intent PURPOSE OR GENERAL IDEA OF BILL: This bill would change "forcible touching" to "unwanted touching" and close a loophole in the state's law by removing the element of intent. SUMMARY OF PROVISIONS: Section one amends section 130.52 of the penal law.