Bill S4440 version original - Nathalia Fernandez
BILL NUMBER: S4440 SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the executive law and the mental hygiene law, in relation to establishing a state director of drug policy PURPOSE OR GENERAL IDEA OF BILL: To establish a State Director of Drug Control Policy. SUMMARY OF PROVISIONS: Section 1 amends the executive law to create a Director of Drug Control Policy and outline their duties.