Bill S4410 version original - Nathalia Fernandez
BILL NUMBER: S4410 SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the election law, in relation to transition and inaugu- ration receipts and expenditures PURPOSE OR GENERAL IDEA OF BILL: To require statements be filed in regards to the receipts and expendi- tures of a Transition and Inauguration entity to increase transparency. SUMMARY OF PROVISIONS: Section 1.