Bill S4460 version original - Shelley Mayer
BILL NUMBER: S4460 SPONSOR: MAYER TITLE OF BILL: An act to amend the tax law, in relation to providing a tax credit for certain taxpayers who suffer the birth of a stillborn child PURPOSE: To provide recognition of the loss and compensation for the costs asso- ciated with a stillbirth. SUMMARY OF SPECIFIC PROVISIONS: Section 1 of the bill amends section 606 of the tax law by adding a new subsection (jjj) which establishes a $2000 refundable tax credit for the birth of a stillborn child.