Bill S4469 version original - Monica Martinez
BILL NUMBER: S4469 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the election law, in relation to the use of school buildings as polling sites PURPOSE: The purpose of this bill is to direct the board or agency in control of a school building to cancel in-person instructional activity when the building is operating as a poll site if a space that is separate or isolated from instructional activities cannot be provided. SUMMARY OF PROVISIONS: Section 1.