Bill S4540 version original - Jessica Ramos
BILL NUMBER: S4540 SPONSOR: RAMOS TITLE OF BILL: An act to amend the labor law, in relation to clarifying that deductions not be made from employee accounts if such deductions do not constitute overpayment or are dispersed at an improper time PURPOSE: Clarifying that deductions from wages are not authorized even if the result of a clerical error or other mistake by the employer if the wages were rightfully owed. SUMMARY OF PROVISIONS: Subdivision c of § 193 of the Labor law § 42-a.