Bill S4497A version original - Andrew Gounardes
BILL NUMBER: S4497 SPONSOR: GOUNARDES TITLE OF BILL: An act to amend the education law, in relation to the license require- ments for the practice of respiratory therapy PURPOSE OR GENERAL IDEA OF BILL: To update the license requirements for respiratory therapy consistent with current national standards. SUMMARY OF PROVISIONS: Section 1: Changes the educational requirement from an associate to a bachelor's degree for licensure as a respiratory therapy.