Bill S4525 version original - Pamela Helming
BILL NUMBER: S4525 SPONSOR: HELMING TITLE OF BILL: An act establishing the veterans' memorials preservation act PURPOSE: This bill seeks to provide statutory protection for veterans' memorials throughout New York State. SUMMARY OF PROVISIONS: Section 1: Cites the title of act to be known as the "Veterans' Memo- rials Preservation Act" Section 2: (a) provides the definition for a memorial; (b) provides the definition for a public entity.