Bill S4524 version original - Pamela Helming
BILL NUMBER: S4524 SPONSOR: HELMING TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to mandatory screening procedures PURPOSE: To provide an additional tool for intervention in cases of defendants who suffer from addiction and would benefit from intervention or a fair clinical assessment of their situation. This amendment would aid the courts in addressing this crisis at a local level.