Bill S4472 version original - Dan Stec
BILL NUMBER: S4472 SPONSOR: STEC TITLE OF BILL: An act to amend the penal law, in relation to adding St. Lawrence county sheriff as an additional firearms licensing officer PURPOSE: To authorize the St. Lawrence County Sheriff to act as an additional firearms licensing officer SUMMARY OF PROVISIONS: Section 1 amends § 265.00(10) of the Penal Law, to allow the Sheriff in St. Lawrence County to act as an additional firearms licensing officer. Section 2 is the effective date.