Bill S4464 version original - Pamela Helming
BILL NUMBER: S4464 SPONSOR: HELMING TITLE OF BILL: An act to amend the penal law, in relation to authorizing the possession of a firearm, shotgun or rifle in certain sport shooting and target practice facilities PURPOSE: This bill authorizes the possession of a firearm, shotgun or rifle in certain sport shooting and target practice facilities. SUMMARY OF PROVISIONS: Amends Section 1. Paragraphs (o) and (p) of subdivision 2 of section 265.