Bill S4538 version original - Thomas F. O'Mara
BILL NUMBER: S4538 SPONSOR: O'MARA TITLE OF BILL: An act to amend the executive law, in relation to establishing the New York state familial search policy and the release of information for familial DNA searches PURPOSE: This bill would authorize familial DNA searching in New York and author- ize the DNA subcommittee to create a report on familial searching. SUMMARY OF PROVISIONS: Section 1.