Bill S4474 version original - Robert Rolison
BILL NUMBER: S4474 SPONSOR: ROLISON TITLE OF BILL: An act to amend the family court act and the criminal procedure law, in relation to orders of protection PURPOSE OR GENERAL IDEA OF BILL:. Enacts "Melanie's law"; relates to orders of protection; allows courts to issue orders of protection for immediate family members or househould members of the named victims, regardless of age. Section 1: Provides for the short title of this act, which shall be known and may be cited as "Melanie's Law.