Bill S4505A version original - John Liu
BILL NUMBER: S4505 SPONSOR: LIU TITLE OF BILL: An act to amend the education law, in relation to prohibiting discrimi- nation of students on school property or at a school function based on religious attire PURPOSE: The intent of this legislation is to further protect the religious rights of all New Yorker's, especially our youth. New York State is a multicultural state made up of many religious denominations, and we must make every effort to pursue equal representation.