Bill S4471 version original - Dan Stec
BILL NUMBER: S4471 SPONSOR: STEC TITLE OF BILL: An act to amend the tax law, in relation to creating a tax credit for instructors of a hunting safety course PURPOSE: To provide a tax credit for instructors of a hunting safety course SUMMARY OF PROVISIONS: Section 1 amends the tax law by adding a new subdivision (ccc) to § 606 establishing a hunting safety course instructor tax credit equal to $300. Section 2 provides the effective date.