Bill S4584 version original - Anthony H. Palumbo
BILL NUMBER: S4584 SPONSOR: PALUMBO TITLE OF BILL: CONCURRENT RESOLUTION OF THE SENATE AND ASSEMBLY proposing an amendment to section 14 of article 3 of the constitution, in relation to limiting the use of messages of necessity PURPOSE: This bill would limit the use of messages of necessity to respond to a state disaster. SUMMARY OF PROVISIONS: Amends section 14 of the state constitution to limit the use of a message of necessity to respond to a state disaster emergency.