Bill S4570 version original - James Skoufis
BILL NUMBER: S4570 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the criminal procedure law, in relation to securing orders for principals charged with arson felony offenses PURPOSE OR GENERAL IDEA OF BILL: To add felony arson in the third degree and the fourth degree to the list of bail eligible offenses SUMMARY OF PROVISIONS: Section 1. This act shall be known as "Billy's Law" Section 2. Adds felony arson in the third and fourth degree to the list of bail eligible offenses Section 3.