Bill S4590 version original - James Sanders Jr.
BILL NUMBER: S4590 SPONSOR: SANDERS TITLE OF BILL: An act to amend the executive law, in relation to reporting duties with respect to missing children and adults PURPOSE: To amend the executive law, in relations to reporting duties with respect to missing children and adults. SUMMARY OF PROVISIONS: Section 1 states this act may be cited as Lamont Dottin's law.