Bill S4559C version original - Monica Martinez
BILL NUMBER: S4559 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the general business law, in relation to enacting the "school security guard training enhancement act" PURPOSE: The purpose of this bill is to equip school security guards with adequate training conducive to mitigating issues that arise in an educa- tional environment pertaining to the youth population. SUMMARY OF PROVISIONS: Section 1.