Bill S4550B version original - Rachel May
BILL NUMBER: S4550 SPONSOR: MAY TITLE OF BILL: An act to amend the education law, in relation to enacting the "schools impacted by gross highways (SIGH) act" PURPOSE: The purpose of this bill is to prohibit the construction of new schools within 500 hundred feet of a controlled-access highway unless space is so limited that there is no other site to erect a school. SUMMARY OF PROVISIONS: Section 1: sets the name. Section 2: adds a new subdivision 3-a to section 408 of the education law.