Bill S4567A version original - Dan Stec
BILL NUMBER: S4567 SPONSOR: STEC TITLE OF BILL: An act to amend the tax law, in relation to extending the authorization granted to the county of St. Lawrence to impose an additional one percent of sales and compensating use taxes PURPOSE: The legislation authorizes St. Lawrence County to continue collecting an additional 1 percent of sales and compensating use tax for a period of three years. SUMMARY OF PROVISIONS: Section 1 - Authorizes St.