Bill S4587 version original - Peter Oberacker
BILL NUMBER: S4587 SPONSOR: OBERACKER TITLE OF BILL: An act to amend the education law, in relation to payment of moneys due for prior years PURPOSE: To utilize prior year claims adjustments to offset state aid adjustments for overpayment of funds to school districts. SUMMARY OF PROVISIONS: Section 3604 of the education law is amended to provide school districts with more timely payments of prior year claims in the event a state aid adjustment is necessary.