Bill S4560 version original - Monica Martinez
BILL NUMBER: S4560 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the tax law, in relation to the public safety surcharge; and to repeal certain provisions of such law relating thereto PURPOSE: The purpose of this bill is to removes the language which requires the to state move public safety surcharge funds into the state general fund. SUMMARY OF PROVISIONS: Section 1. amends section 325 of the county law by repealing section 5(a). Section 2.