Bill S4574 version original - Jessica Ramos
BILL NUMBER: S4574 SPONSOR: RAMOS TITLE OF BILL: An act to amend the executive law, in relation to ensuring equal access to health and human services for limited English speaking individuals PURPOSE: The purpose of this legislation is to amend workers compensation law so that irregularities in a cover sheet accompanying an application for full board review shall not be grounds for denial of said application for administrative review or full board review. SUMMARY OF PROVISIONS: Section 1.