Bill S4554 version original - Pamela Helming
BILL NUMBER: S4554 SPONSOR: HELMING TITLE OF BILL: An act to amend the state technology law and the tax law, in relation to a refundable tax credit provided to broadband deployment to residences and small businesses located in rural, unserved areas PURPOSE: This bill promotes rural broadband deployment by providing residents and small businesses with tax credits for broadband deployment. SUMMARY OF PROVISIONS: Section 1: Sites the name of this act as the Credit for Rural Broadband Act Section 2.