Bill S4542 version original - Steve Rhoads
BILL NUMBER: S4542 SPONSOR: RHOADS TITLE OF BILL: An act to amend the election law, in relation to publicly releasing ballot images PURPOSE OR GENERAL IDEA OF BILL: To require the Board of Elections to make the images of ballots avail- able for public scrutiny SUMMARY OF PROVISIONS: Section 1. Amends the election law by adding a new section 7-131 to authorize the Board of Elections to publicly release ballot images following an election. Section 2. Establishes effective date.