Bill S4591 version original - Peter Oberacker
BILL NUMBER: S4591 SPONSOR: OBERACKER TITLE OF BILL: An act to amend the criminal procedure law, in relation to the use of video monitoring equipment to conduct arraignments SUMMARY OF PROVISIONS: Section 1 amends article 182.20 of the Criminal Procedure Law expanding the use of virtual arraignments to any county in the State. Section 2 provides for the effective date.