Bill S4589 version original - James Sanders Jr.
BILL NUMBER: S4589 SPONSOR: SANDERS TITLE OF BILL: An act to amend the public health law, in relation to adding dysmenor- rhea to the list of conditions covered for lawful medical use of mari- huana PURPOSE: This bill adds dysmenorrhea to the list of conditions for which a doctor may authorize medical marijuana use. SUMMARY OF SPECIFIC PROVISIONS: Section one amends subparagraph (i) of paragraph (a) of subdivision 7 of section 3360 of the public health law.