Bill S4596 version original - Timothy M. Kennedy
BILL NUMBER: S4596 SPONSOR: KENNEDY TITLE OF BILL: An act to amend the local finance law, in relation to the sale of munic- ipal obligations by the county of Erie PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to provide Erie County with flexibility in its bond sales to maximize its return on bonds sold. SUMMARY OF PROVISIONS: Section 1 amends section 54.