Bill S4566 version original - Kevin S. Parker
BILL NUMBER: S4566 SPONSOR: PARKER TITLE OF BILL: An act to maintain the continued viability of the state's existing biom- ass facility which is located within a military installation PURPOSE OF THE BILL: To allow for the continued operation of a biomass facility located on a military installation and providing 100% of the military installation's electricity. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Legislative Intent.