Bill S4621 version original - Luis R. Sepúlveda
BILL NUMBER: S4621 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the correction law, in relation to incarcerated individ- uals with a serious mental illness PURPOSE: This bill changes the definition of serious mental illness in the correction law to match the definition in the mental hygiene law. SUMMARY OF PROVISIONS: Section 1 amends subparagraph i of paragraph e of subdivision 6 of section 137 of the correction law. Section 2 provides an effective date.