Bill S4620 version original - Joseph A. Griffo
BILL NUMBER: S4620 SPONSOR: GRIFFO TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to restricting driving in a passing lane without overtaking or passing a vehicle PURPOSE: The purpose of this bill is to prevent drivers from occupying the left lane of an interstate highway for reasons other than passing or overtak- ing other vehicles. SUMMARY OF PROVISIONS: Section 1 sets the title of the bill.