Bill S4683 version original - Kevin S. Parker
BILL NUMBER: S4683 SPONSOR: PARKER TITLE OF BILL: An act to amend the public health law, in relation to the commercial display of human remains PURPOSE: To require a permit issued by the Health Department if human remains are to be displayed for a commercial purpose such as a museum or art exhib- its. SUMMARY OF PROVISIONS: Section 1: Legislative findings. Section 2: The public health law is amended by adding a new article 43-C.