Bill S4711 version original - Monica Martinez
BILL NUMBER: S4711 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the volunteer firefighters' benefit law and the volun- teer ambulance workers' benefit law, in relation to the payment of death benefits PURPOSE: To establish the payment of certain volunteer firefighter and volunteer ambulance worker death benefits to be paid within ninety days of the filing of application to receive such death benefits. SUMMARY OF PROVISIONS: Section 1.