Bill S4665 version original - Andrew J. Lanza
BILL NUMBER: S4665 SPONSOR: LANZA TITLE OF BILL: An act to amend the criminal procedure law, in relation to hate crimes and bail eligibility PURPOSE: Relates to hate crimes and bail eligibility. SUMMARY OF PROVISIONS: Section one amends paragraphs (s) and (t) of subdivision 4 of section 510.10 of the criminal procedure law, as added by section 2 of part UU of chapter 56 of the laws of 2020, are amended and a new paragraph (u). is added.