Bill S4661 version original - Andrew J. Lanza
BILL NUMBER: S4661 SPONSOR: LANZA TITLE OF BILL: An act to amend the public authorities law, in relation to special powers of the metropolitan transportation authority PURPOSE OR GENERAL IDEA OF BILL: To provide easier access to paratransit for eligible individuals. SUMMARY OF SPECIFIC PROVISIONS: This legislation amends Section 1266 of the public authorities law by adding subdivision 15-a. JUSTIFICATION: Filling out current forms required to apply for paratransit may be difficult for seniors.