Bill S4617 version original - Joseph A. Griffo
BILL NUMBER: S4617 SPONSOR: GRIFFO TITLE OF BILL: An act to amend the general business law, in relation to raising the maximum fine for persons who violate the law regulating telemarketing PURPOSE: The purpose of this bill is to deter any would-be violators of the Do Not Call Registry. SUMMARY OF PROVISIONS: Section 1 amends the general business law by increasing the fine applied to violators of the Do Not Call Registry to $20,000. Section 2 sets the effective date.