Bill S4631 version original - Andrew J. Lanza
BILL NUMBER: S4631 SPONSOR: LANZA TITLE OF BILL: An act to amend the education law, in relation to including the Sandy Ground community in the education curriculum PURPOSE: Includes the Sandy Ground community in the education curriculum. SUMMARY OF PROVISIONS: Section 1. Amends parts of the education law to include the history of the Sandy Ground community, the oldest continuously inhabited free black settlement in the United States, in the curriculum. Section 2.