Bill S4714 version original - Rachel May
BILL NUMBER: S4714 SPONSOR: MAY TITLE OF BILL: An act to amend the not-for-profit corporation law, in relation to exempting the Southwood Volunteer Fire Department, Inc. from the requirement that the percentage of non-resident fire department members not exceed forty-five percent of the membership PURPOSE: To authorize non-resident volunteer membership within the Southwood Volunteer Fire Department, Inc.