Bill S4610 version original - Andrew J. Lanza
BILL NUMBER: S4610 SPONSOR: LANZA TITLE OF BILL: An act to amend the penal law, in relation to arson in the first degree PURPOSE: This bill expands and clarifies the definition of arson in the first degree to include fires accelerate by various incendiary materials, and fires caused or accelerated by the use of flammable or combustible liquids or flammable gases. SUMMARY OF PROVISIONS: This bill would expand and clarify the definition of arson in the first degree, section 150.