Bill S4686A version original - Kevin S. Parker
BILL NUMBER: S4686 SPONSOR: PARKER TITLE OF BILL: An act to amend the criminal procedure law and the executive law, in relation to enacting the "New York State Phoenix Act" PURPOSE: The purpose of the bill would be to establish the Phoenix Act and crime of Coercive Control. SUMMARY OF PROVISIONS: Section 1. Establishes that the act shall be known as the Phoenix Act. Section 2. Legislative findings. Section 3. Subdivision 3 of section 30.