Bill S4709A version original - Brian Kavanagh
BILL NUMBER: S4709 SPONSOR: KAVANAGH TITLE OF BILL: An act to amend the real property tax law, in relation to authorizing a tax exemption and a tax abatement for alterations and improvements to multiple dwellings for purposes of preserving habitability in affordable housing PURPOSE: Section 489 of the Real Property Tax Law ("RPTL"), known in the City of New York as the J-51 Program, expired on June 29, 2022. This bill amends and extends the J-51 Program.