Bill S4657 version original - Andrew J. Lanza
BILL NUMBER: S4657 SPONSOR: LANZA TITLE OF BILL: An act to amend the penal law, the general business law and the vehicle and traffic law, in relation to criminal use of a firearm; and to repeal sections 265.08 and 265.09 of the penal law relating thereto PURPOSE: Establishes class A-1 felony of criminal use of a firearm for possessing a deadly weapon or displaying a gun during the commission of a class A or violent felony. SUMMARY OF PROVISIONS: Section 1.