Bill S4616 version original - Joseph A. Griffo
BILL NUMBER: S4616 SPONSOR: GRIFFO TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to the weight of certain trailers and semi-trailers driven on public highways PURPOSE: To increase the allowable weight of trailers and semi-trailers from 1,000 pounds unladen to 1,500 pounds unladen. SUMMARY OF PROVISIONS: Section 1 amends the vehicle and traffic law to increase the allowable weight of trailers and semi-trailers from 1,000 pounds unladen to 1,500 pounds unladen.